+++ src/Symfony/Bundle/FrameworkBundle/Tests/Command/KernelInterface.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\Command;
+
+use Symfony\Component\HttpKernel\KernelInterface as BaseKernelInterface;
+
+interface KernelInterface extends BaseKernelInterface
+{
+    public function getMode(): string;
+}