+++ src/Symfony/Component/HttpKernel/KernelInterface.php
@@ -23,6 +23,7 @@ use Symfony\Component\HttpKernel\Bundle\BundleInterface;
  * @method string getBuildDir() Returns the build directory - not implementing it is deprecated since Symfony 5.2.
  *                              This directory should be used to store build artifacts, and can be read-only at runtime.
  *                              Caches written at runtime should be stored in the "cache directory" ({@see KernelInterface::getCacheDir()}).
+ * @method string getMode()
  *
  * @author Fabien Potencier <fabien@symfony.com>
  */
@@ -91,6 +92,8 @@ interface KernelInterface extends HttpKernelInterface
      * Gets the environment.
      *
      * @return string The current environment
+     *
+     * @deprecated since Symfony 5.2, use getMode() instead.
      */
     public function getEnvironment();